Enhanced greedy coarsener.
It does not replicate every edge during coarsening. It contracts
edges, using the weight property of the edge to store how many edges has
been pulled together. The greedy node selection prefers nodes that did
not participate in the previous pass of the coarsening. Unfortunately
this model still act as a black hole for evenly distributed graphs.
It produces one massive node and a few small ones.