Added some extra logging to BucketPartitioner.
BUG: BucketPartitioner cannot handle a set of nodes, when the first is
a really big supernode (bigger than the allowed size of the partitions).
There is a test case that highlights the issue. Possible solution could be
that we split up the one big node, so it can fit into our partitions, but
actually that would mix up concenrs in this class. Let this be last resort.